update
Took 2 hours 58 minutes

+++ bulk/TransferTool.sol
@@ -39,7 +39,7 @@ contract TransferTool {
         return true;
     }
 
-    function checkBalance() public view returns (uint) {
+    function checkBalance() public view returns (uint256) {
         return address(this).balance;
     }
